Pricing Changes

+++ Models/pricingModel.js
@@ -1,6 +1,6 @@
 module.exports = (sequelize, DataTypes) => {
     const Pricing = sequelize.define("Pricing", {
-      costperhr: { type: DataTypes.INTEGER },
+      costperhr: { type: DataTypes.FLOAT },
       carid: { type: DataTypes.INTEGER, primaryKey: true },
     });
     return Pricing;
